Rename Makefile     --> Makefile.in Rename csprof       --> csprof.in Rename csprof_syms  --> csprof_syms.in
All of these are renamed to their '.in' equivalent so that the same directory can
be used for build and install.

Added config.dev to generate a Makefile (for TAGS) in the source directory
NOTE: this will *NOT* work if the build directory = source directory.

config.dev is a config script for developer convenience. Most users will not need this.