only filter local named symbols, leave global named symbols unfiltered.
reason: in libc-2.7.so, the function __memset_chk tailcalls __chk_fail with a conditional jump
this caused csprof_syms to elide the function memset (among others) because they fell in the span of the
conditional jump. To avoid this, we fine-tuned the filtering to only filter symbols found in
the symbol table if they are not global symbols. memset is a global symbol.

0000003ec167a880 <__memset_chk>:
__memset_chk():
  3ec167a880:	48 39 d1             	cmp    %rdx,%rcx
  3ec167a883:	0f 82 f7 e1 06 00    	jb     3ec16e8a80 <__chk_fail>
  3ec167a889:	0f 1f 80 00 00 00 00 	nopl   0x0(%rax)

0000003ec167a890 <memset>:
memset():
  3ec167a890:	48 83 fa 07          	cmp    $0x7,%rdx
  3ec167a894:	48 89 f9