e the host wordsize (32 or 64) to set the multilib search path. This only affects outside packages via --with-pkg (ie, not buit in externals).
Settle on --enable-wordsize instead of --enable-bits.  This option is
almost never needed and could just go away.

Add sanity checks that CC and CXX generate the same wordsize.  Also
warn if -m32 or -m64 is used in CFLAGS and not in CC or CXX.  This
shouldn't be necessary, some packages can't handle -m32 in CFLAGS and
need CC='gcc -m32'.

Fix a couple obscure bugs related to specifying outside packages.