hpcrun: Resolve the final [!] bug normalizing IPs.
hpcrun_addr_to_interval(): Extend semantics to support normalization
of the correct IP.

Background: To support fast IP normalization, we memoized load module
information in the unwind-interval splay tree and then extended the
semantics of hpcrun_addr_to_interval() to return a normalized IP using
the memoized load module information.

However: There are cases in the x86 unwinder where
hpcrun_addr_to_interval() is called not with the official IP, but with
IP plus/minus some offset.  This, of coarse, means that the normalized
IP recorded in the unwind cursor is in error by a small amount.  The
above change resolves this issue.