Support a separate build directory by copying the package skeleton directories into the build tree and symlinking distfiles.  This is essentially what automake would do.
Allow the user to set the prefix and copy externs.conf into the prefix
directory.  Now, the --with-externals=DIR option to hpctoolkit can be
either the build or prefix directory.

This allows building two copies of externals (into two prefixes) from
the same source tree, for example for front and back end nodes.