Protect hpcrun's tracing routines with an "hpcrun_" prefix.  Now trace routines have the form:   hpcrun_trace_*
Motivation (other than general spring cleaning): I just tried using
hpclink with a code that happens to contain a utility call to
trace_init(), which clashes with HPCToolkit's name.  The only thing
that saved us during the link was the fact that, being in a Fortran
source file, the symbol turned into "trace_init_".