Install plugin config files for the GA, IO and MEMLEAK sample sources. For now, this only applies to hpclink and the static case.  For example, for memleak, the following two command lines are equivalent.
   hpclink --memleak gcc file.c ...
   hpclink --plugin memleak gcc file.c ...

As we move forward, we'll phase out the individual option (--io) for
each source in favor of the more general plugin method (--plugin io).
For hpclink, extend the plugins to read the variable

   hpclink_undefined_names='name ...'

for a list of symbol names to force as undefined.  I'm not sure if
this useful or needed (I think it doesn't really matter), but this is
what hpclink currently does.  So, we'll support the variable and let
the plugin decide what it wants.

This commit just copies the plugin config files into the install tree
Building the libhpcrun_plugin_wrap.a files is another matter.  That
probably still needs to be done in the hpcrun Makefile.am to pull in
all its mess of platform-specific cflags, etc.

Also, refix the wrap names for the IO source.  I accidentally dropped
read and write in an earlier commit, probably due to a careless merge
on my part.