We attempt to connect method_id with address. In compiled java byte code (CompiledMethodLoad callback) we can identify a method id and its address. However, this is not the case for other callbacks such as DynamicCodeGenerated where we don't have any information about method_id.
At the moment we just assume a method id is persistent. In reality this is not the case, since JVM can unload classes and a method id is assigned to a new address.

We are using splay tree to store the <method_id, address> pair. See java/jmt.c
Unfortunately, hpcrun Splay tree is designed for interval node, so we need to have our own "regular" splay tree (at the moment is located in java/splay-general.h)