Finish John's commit to the html pages and remove some commented-out links to Scidac Outreach.
Note: when editing the html pages, edit the html.in files and then
rerun make (from the build directory) to regen the .html files.  The
safest procedure is: edit the html.in files, 'touch *.in', then make
and make install.

Note: view the results and copy to the web site from the install
directory.