hpcstruct: add an option '--use-binutils' to enable hpcstruct to use binutils decoder to crack binary instructions
Implementation notes:
We have now 2 new classes derived from x86ISA: x86ISAXed and x86ISABinutils.
We check if we use binutils or not inside LM::open() method.
The LM object will use x86ISAXed instance if m_useBinutils attribute is true, otherwise x86ISABinutils class in instantiated:

230   ISA* newisa = NULL;
...
247       if (m_useBinutils) {
248          newisa = new x86ISABinutils(true);
249       } else {
250          newisa = new x86ISAXed(true);
251       }

The m_useBinutils attribute is set during the constructor instead of during opening the file. 
The reason is we don't want to have part of structure is decoded with Xed, and the other part with binutils. It's either using Xed, or Binutils.


General notes:
- on x86 the default will be Xed (unless the --use-binutils flag is set)
- on other platforms the default will be binutils. '--use-binutils' flag will be useless.
- not tested yet on non-x86 platforms