Adjust the quoting rules in externals's Makefile.inc.  Use single quotes for the CMD variables in Makefile.inc.  This allows using double quotes in the commands in the package Makefiles.
For Xed on k1om, add the libirc.a file in a post-fetch command instead
of the install script.  This is a cleaner way of adding the libirc
library.  In general, we want to unpack all of the tar files into the
local work directory at the make fetch step.