Instead of using method entry/exit to store call stacks, we'll use the undocumented AsyncGetCallTraces
There's a working example of using AsyncGetCallTraces in an open source lightweight java profiler: 

https://code.google.com/p/lightweight-java-profiler

It's known that AsyncGetCallTraces doesn't work in many cases, and sometimes the information of  method IDs or Java class doesn't even exist (it all returns 0 in the frames for most cases).

However, the trick is to force JVM to generate method IDs and classes during "class prepare" event. 
In this event we should ask JVM to generate method IDs and classes, and then just throw them away.
There's penalty for this, but still much less than overhead using method entry/exit.