Finish updating symtabAPI to 8.2 with inline support.  This is a pre-8.2 dyninst snapsot from git.
Symtab 8.2 now requires cmake, GNU gcc/g++ 4.4 or later and support
for std C++11.  Configure tests for these and will warn or fail if not
available.

The config option '--disable-symtab82' disables building 8.2 and
reverts to the previous symtab 8.0.

The config var 'CXX11_FLAG' specifies a CXX flag for std C++11 in case
configure fails to find a suitable flag.  This is really only needed
for a non-GNU C++ compiler.