Add a second procedure name to the Alien class.  The old procnm (m_name) is used to locate nodes in the tree, including the fake 'inline-alien-xxx' names.  The new displaynm is output in the 'n=' field of the Alien node in the hpcstruct file.
Also, remove the 'tosOnCreate' parameter from demandAlienStrct().
None of the calls use 'true' anymore.

Although the fake 'inline-alien-xxx' proc names in the AlienStrctMap
cache seem to work and find the right nodes in a sequence of
intermediate aliens, we may want to rethink the cache map.
Specifically, we may want to separate the intermediate aliens from
the guard aliens and treat them differently.