hpcprof: allow redundancy for some specific files such as "unknown-file" and empty file (somehow alien procedures do have empty files).
Ideally, the file table in experiment.xml contains the absolute file path. When hpcprof fails to find the exact filename, it names to "~unknown-file" or for alien cases it's an empty name (not sure why). 

Below is an example of the same file (unknown-file) which came from different load modules (2 and 208997).

    <LoadModule i="2" n="/home/la5/pkgs/hpctoolkit/libexec/hpctoolkit/hpcstruct-bin"/>
    <LoadModule i="208997" n="/home/projects/pkgs/gcc-4.6.3/lib64/libstdc++.so.6.0.16"/>

    <File i="175060" n="~unknown-file~"/>
    <File i="209018" n="~unknown-file~"/>

    <File i="175693" n=""/>
    <File i="175718" n=""/>

..
<PF i="504" s="176000" l="0" lm="2" f="175060" n="176000">
...
<PF i="598" s="209019" l="0" lm="208997" f="209018" n="209019">