Add zlib to externals, hpcserver will use this.  But for now, the default is off.
Welcome to Mark's 95% hell.  Almost all Linux systems have zlib, but
I've found two problems.  (1) some systems (eg, guadalupe in RCSG)
don't have the zlib devel package installed, so there is no zlib.h.
(2) the back-end MPI compiler on some systems (eg, mira at ANL)
doesn't use /usr/include in their -I paths, so again, there is no
zlib.h.

Zlib is small and builds easily, so we'll just cave and build it in
externals.  It's either that or add directions on how to download and
build zlib for the 5% of systems that need it.  At least it's an
actual library, not infrastructure like cmake.  Blech.