Fix the 'white trace line' glitch in the trace server.  There was a bug in the alignment of the memory chunks where it was possible that one trace record could be split across two segments.  This resulted in the low-order bits of the time stamp being bogus for a very few trace records that crossed a segment boundary.
The workaround I've chosen is to extend each segment by a few bytes so
that no trace record is split across two segments.  This duplicates a
few records that would otherwise cross a segment boundary.  (Fine as
long as the file is read-only.)

This may seem like a hack, but actually it's not so much.  We already
round up the chunk size to the mmap page size, so a few extra bytes is
not a problem.  And it greatly simplifies handling this corner case.