Fix some usage that standard C++11 barfs on.  Two uses of note:
1. C++11 does not allow 'const' for non-integer constants (float,
double) inside a class.  For example:

class Foo {
public:
   static const double epsilon = 0.000001;
}

In C++11, this must be 'constexpr'.  Apparently 'const' is not
constant enough. </sarcasm>  Something to do with arithmetic at compile
time versus run time not being portable for floating point.  But
'constexpr' is not supported without -std=c++11, so there is no usage
that works for both C++11 and non-C++11.

In this case, our use of 'epsilon' is small, so we'll just use #define
until we adopt C++11 and can use constexpr.

2. C++11 requires a space between adjacent quotes (") for concatenating
string constants.  This often occurs with the printing macros for the
standard int types in <inttypes.h>.  For example:

uint64_t x;
printf("x = %"PRIu64"\n", x);

C++11 requires a space around 'PRIu64' since PRIu64 expands to "lu".
This one is just a warning.

At this point, toolkit builds with or without -std=c++11.  At some
point, we'll need to adopt -std=c++11, but not yet.